refactor: reorganize custom LLM service into modular services directory
- Extract service state management into services/custom_llm_state.py
- Move chat models and streaming logic to dedicated modules
- Simplify EmotionAgent initialization with environment variable fallbacks
- Remove LLM client abstraction complexity in favor of direct implementation
- Improve error handling and logging throughout services
