+++ app/services/__init__.py
@@ -1,0 +1,3 @@
+"""
+Service layer helpers for the custom LLM FastAPI service.
+"""
